+++ Mods/lx_enhanced_divine_combat_3ff156e2-289e-4dac-81f5-a44e3e304163/Story/RawFiles/Lua/BootstrapServer.lua
@@ -6,7 +6,7 @@ Ext.Require("LXDGM_Potions.lua")
 Ext.Require("LXDGM_Talents.lua")
 Ext.Require("LXDGM_Weapons.lua")
 Ext.Require("LXDGM_StatsPatching.lua")
-Ext.Require("LXDGM_AttributeBonuses.lua")
+Ext.Require("LXDGM_CustomBonuses.lua")
 
 ---- General Functions ----
 ---@param character EsvCharacter