+++ Mods/lx_enhanced_divine_combat_3ff156e2-289e-4dac-81f5-a44e3e304163/Story/RawFiles/Lua/LXDGM_DamageControl.lua
@@ -279,7 +279,6 @@ function DGM_CalculateHitChance(target, attacker)
 	local dodge = target.Dodge
 
 	local chanceToHit1 = accuracy - dodge
-	Ext.Print(accuracy, dodge, chanceToHit1)
 	chanceToHit1 = math.max(0, math.min(100, chanceToHit1))
     return chanceToHit1 + attacker.ChanceToHitBoost
 end